+++ Core/Graphics/Image.cs
@@ -110,12 +110,15 @@ public class Image
         return new(indices, dimension, ImageType.Palette, offset, ns);
     }
 
-    public static Image? FromImageSharp<TPixel>(SixLabors.ImageSharp.Image<TPixel> data, Vec2I imageOffset = default, ResourceNamespace ns = ResourceNamespace.Global)
+    public static Image? FromImageSharp<TPixel>(SixLabors.ImageSharp.Image<TPixel> data, Vec2I imageOffset = default, ResourceNamespace ns = ResourceNamespace.Global,
+        Palette? palette = null)
         where TPixel : unmanaged, IPixel<TPixel>
     {
-        byte[] argbData = new byte[data.Height * data.Width * 4];
+        var indices = palette == null ? null : new ushort[data.Height * data.Width];
+        var argbData = new byte[data.Height * data.Width * 4];
         int offset = 0;
         Rgba32 tempPixel = new();
+        int index = 0;
         for (int y = 0; y < data.Height; y++)
         {
             Span<TPixel> pixelRow = SixLabors.ImageSharp.Advanced.AdvancedImageExtensions.DangerousGetPixelRowMemory(data, y).Span;
@@ -127,14 +130,19 @@ public class Image
                 argbData[offset + 1] = tempPixel.R;
                 argbData[offset + 2] = tempPixel.G;
                 argbData[offset + 3] = tempPixel.B;
+
+                if (indices != null && palette != null && tempPixel.A != 0)
+                    indices[index] = palette.GetNearestColorIndex(new(tempPixel.R, tempPixel.G, tempPixel.B));
+
+                index++;
                 offset += 4;
             }
         }
 
-        return FromArgbBytes((data.Width, data.Height), argbData, imageOffset, ns);
+        return FromArgbBytes((data.Width, data.Height), argbData, imageOffset, ns, indices);
     }
 
-    public static Image? FromArgbBytes(Dimension dimension, byte[] argbData, Vec2I offset = default, ResourceNamespace ns = ResourceNamespace.Global)
+    public static Image? FromArgbBytes(Dimension dimension, byte[] argbData, Vec2I offset = default, ResourceNamespace ns = ResourceNamespace.Global, ushort[]? indices = null)
     {
         if (dimension.Area * 4 != argbData.Length)
             return null;
@@ -155,7 +163,8 @@ public class Image
             argbByteOffset += 4;
         }
 
-        return new(pixels, dimension, ImageType.Argb, offset, ns);
+        var imageType = indices == null ? ImageType.Rgba : ImageType.PaletteWithArgb;
+        return new(pixels, dimension, imageType, offset, ns, indices);
     }
 
     public Image GetUpscaled(int upscalingFactor)
@@ -355,11 +364,11 @@ public class Image
             m_indices[offset] = index;
     }
 
-    public void SetPixel(int x, int y, Color color, Colormap colormap)
+    public void SetPixel(int x, int y, Color color, Palette palette)
     {
         byte index = 0;
         if (ImageType == ImageType.PaletteWithArgb)
-            index = colormap.GetNearestColorIndex(color);
+            index = palette.GetNearestColorIndex(color);
 
         int offset = (y * Width) + x;
         if (offset >= 0 && offset < m_pixels.Length)
@@ -450,11 +459,11 @@ public class Image
 
         for (int y = 0; y < HalfDimension; y++)
             for (int x = 0; x < HalfDimension; x++)
-                image.SetPixel(x, y, Color.Red, Colormap.GetDefaultColormap());
+                image.SetPixel(x, y, Color.Red, Palette.GetDefaultPalette());
 
         for (int y = HalfDimension; y < Dimension; y++)
             for (int x = HalfDimension; x < Dimension; x++)
-                image.SetPixel(x, y, Color.Red, Colormap.GetDefaultColormap());
+                image.SetPixel(x, y, Color.Red, Palette.GetDefaultPalette());
 
         return image;
     }